pyproject.toml
[tool.poetry]
name = "monument ai"
version = "0.1.0"
description = ""
authors = ["Zaskilovan <diulgervanya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
tortoise-orm = "^0.20.0"
fastapi-pagination = "^0.12.22"
mypy = "^1.9.0"
black = "^24.3.0"
isort = "^5.13.2"
aerich = "^0.7.2"
flake8 = "^7.0.0"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
mysql = "^0.0.3"
aiomysql = "^0.2.0"
cryptography = "^42.0.5"



[tool.aerich]
tortoise_orm = "config.database_config"
location = "../migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

